requirements.txt
coinbase-advanced-py>=1.7
python-dotenv>=1.0
